requirements.txt
absl-py==0.9.0
appnope==0.1.0
astor==0.8.1
atari-py==0.2.6
attrs==19.3.0
backcall==0.2.0
backports.csv==1.0.7
beautifulsoup4==4.9.1
bleach==3.1.5
blis==0.4.1
boto==2.49.0
boto3==1.14.5
botocore==1.17.5
box2d-py==2.3.8
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cheroot==8.3.0
CherryPy==18.6.0
cld2-cffi==0.1.4
click==7.1.2
cloudpickle==1.2.1
colour==0.1.5
configparser==5.0.0
cycler==0.10.0
cymem==2.0.3
Cython==0.29.20
cytoolz==0.10.1
datasketch==1.5.1
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
feedparser==5.2.1
Flask==1.1.2
Flask-Cors==3.0.8
fr-core-news-sm==2.2.5
funcsigs==1.0.2
future==0.18.2
gast==0.3.3
gensim==3.8.3
glfw==1.11.2
graphviz==0.14
grpcio==1.29.0
gym==0.17.2
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.6.1
importlib-resources==2.0.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.text==3.2.0
jedi==0.17.1
jellyfish==0.8.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.4
jupyterlab-server==1.1.5
kiwisolver==1.2.0
lockfile==0.12.2
lxml==4.3.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
more-itertools==8.4.0
moviepy==1.0.3
msgpack==0.6.2
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nltk==3.5
notebook==6.0.3
numpy==1.18.5
objgraph==3.4.1
opencv-python==4.2.0.34
packaging==20.4
paddlepaddle==1.8.2
pandocfilters==1.4.2
parl==1.3.1
parso==0.7.0
pathlib==1.0.1
pdfminer.six==20200517
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plac==1.1.3
-e git+https://github.com/ntasfi/PyGame-Learning-Environment.git@3dbe79dc0c35559bb441b9359948aabf9bb3d331#egg=ple
portend==2.6
preshed==3.0.2
prettytable==0.7.2
proglog==0.1.9
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.0
ptyprocess==0.6.0
pyarrow==0.13.0
pycparser==2.20
pycryptodome==3.9.7
pyemd==0.5.1
pygame==1.9.4
pyglet==1.5.0
Pygments==2.6.1
pyparsing==2.4.7
Pyphen==0.9.5
pyrsistent==0.16.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.1
PyYAML==5.3.1
pyzmq==18.0.1
rarfile==3.1
redis==3.4.1
regex==2020.6.8
requests==2.23.0
rlschool==0.3.1
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.3.1
Send2Trash==1.5.0
six==1.15.0
smart-open==2.0.0
sortedcontainers==2.2.2
soupsieve==2.0.1
spacy==2.2.4
srsly==1.0.2
tb-nightly==1.15.0a20190801
tempora==3.0.0
tensorboardX==1.8
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
textacy==0.9.1
textpipe==0.11.10
textpipe-pattern==3.6.1
thinc==7.4.0
threadpoolctl==2.1.0
toolz==0.10.0
tornado==6.0.4
tqdm==4.41.1
traitlets==4.3.3
trimesh==3.7.0
urllib3==1.25.9
wasabi==0.6.0
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
zc.lockfile==2.0
zipp==3.1.0
